refactor(test): moved all test files to /test
